feat: add function strems
