+++ input.txt
@@ -1,0 +1,2 @@
+qwedcfsd345hопол
+sdfgsd
