+++ output.txt
@@ -1,0 +1,2 @@
+gmutsvit345xопол
+itvwit
