+++ tools/codec.js
